+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Fixed concatenation support of strings with double numbers
   [#1893](https://github.com/phalcon/zephir/issues/1893)
+- Fixed 'void' return type hint being ignored
+  [#1908](https://github.com/phalcon/zephir/issues/1908)
 
 ## [0.12.2] - 2019-08-05
 ### Added